Refactored; focused and cleaner responsibilites
