+++ src/dotnet-test-mspec/Program.cs
@@ -5,11 +5,6 @@ using Machine.Specifications.Runner.DotNet.Execution.Console;
 using Machine.Specifications.Runner.DotNet.Helpers;
 using Machine.Specifications.Runner.DotNet.Controller;
 using Machine.Specifications.Runner.DotNet.Execution;
-using System.Collections.Generic;
-using Machine.Specifications.Runner.DotNet.Controller.Model;
-using System.Xml.Linq;
-using System.Xml.XPath;
-using System.Linq;
 
 namespace Machine.Specifications.Runner.DotNet
 {
@@ -42,37 +37,17 @@ namespace Machine.Specifications.Runner.DotNet
 
             
 
-            TestController testController = new TestController(mspecAssembly, allListeneres);
+            var testController = new TestController(mspecAssembly, allListeneres);
             if (commandLine.List) {
                 Console.WriteLine(testController.DiscoverTestsRaw(testAssembly));
             } else {
+                var runManager = new RunManager(testController);
                 var assembliesToRun = new[] { testAssembly };
-                var contexts = GetAllContextsFor(testController, assembliesToRun);
-
-                if( commandLine.HasSpecificTestsToRun )
-                    SpecificAssertionsHelper
-                        .GetSpecificationsToRunPerAssemblyFromTestStrings(contexts, assembliesToRun, commandLine.Tests)
-                            .Each(t=>testController.RunMembers(t.Assembly, t.Members));
-                else testController.RunAssemblies(assembliesToRun);
+                runManager.Run(commandLine, assembliesToRun);
 
                 if (runListener.FailureOccurred)
                     Environment.Exit(-1);
             }
-        }
-
-        static IEnumerable<ContextInfo> GetAllContextsFor(TestController testController, IEnumerable<Assembly> assemblies)
-        {
-            var contexts = new List<ContextInfo>();
-
-            foreach( var assembly in assemblies )
-            {
-                var contextsAsString = testController.DiscoverTestsRaw(assembly);
-                var contextsXml = XDocument.Parse(contextsAsString);
-                var contextsElements = contextsXml.XPathSelectElements("/contexts/contextinfo");
-                contexts.AddRange(contextsElements.Select(e=>ContextInfo.GetFrom(e)));
-            }
-
-            return contexts;
         }
 
         private static void PrintVersionInfo(Assembly assembly)
